[internal] Changes related to forked TPCircularBuffer

+++ src/audio/capture_audio/macos/MacAudioInput.h
@@ -13,7 +13,7 @@ class MacAudioInput final: public IAudioInput
 public:
     MacAudioInput(
             AVCaptureDevice *capture_device,
-            int              channels,
+            uint8_t          channels,
             uint32_t         sample_rate,
             uint32_t         frame_size);
